Uses variable path for the directory of all the snow data.
Fuctions at the start are defined are used in python mapping. For
example the function, get_filename(x) will get the basename for the file
for all elements within the list, Files. This is done through the map(lambda x:)
function.

Data can be selected either by a particular month, e.g all march months.
Or the data can be selected by a range using parameters: start and end date.

A colormap is created so that both the flag values [-30,-20, -10, 0] and
the actual data [0-->500] can be plotted on the same graph. This needs work.

final function plots the monthly values (xarray calculates a mean for each month).